implemented siv methods and core analyzer, implemented use cases for Room usage

+++ app/src/main/java/com/example/meditationbiorefactoring/feature_bio/domain/model/MeasurementResult.kt
@@ -1,7 +1,7 @@
 package com.example.meditationbiorefactoring.feature_bio.domain.model
 
 sealed class MeasurementResult {
-    data class Success(val value: Int) : MeasurementResult()
+    data class Success(val value: Double) : MeasurementResult()
     data object Invalid : MeasurementResult()
     data object Error : MeasurementResult()
 }
